Extract Bollinger Bands features

step_bbands creates a specification of a recipe step that will extract Bollinger Bands features from an asset price historical data.

step_bbands(recipe, ..., ma_fun = TTR::SMA, n = 20, sd_mult = 2,
  ma_options = list(), state = FALSE, previous = TRUE,
  state_options = list(high = 1, medhigh = 0.75, medlow = 0.25, low = 0),
  prefix = "bbands", h = NULL, l = NULL, c = NULL, type = NULL,
  role = "predictor", trained = FALSE, skip = FALSE,
  id = rand_id("bbands"))

# S3 method for step_bbands
tidy(x, info = "terms", ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
Either three or one (unquoted) column name(s). If three columns are given, it will represent the "high", "low", and "close" prices, respectively. Otherwise, if only one column name is given, it will treated as "close" price.

ma_fun
A function to extract moving average, or a character vector of length one which specify a moving average function. Defaults to TTR::SMA.

n
A numeric vector of length one which specify the moving average window. The default is 20.

sd_mult
A numeric vector of length one which specify standard deviation multiplier. The default is 2.

ma_options
A list of additional argument(s) that would be passed to ma_fun function.

state
An option to specify whether to return the current states of the Bollinger Bands. Defaults to TRUE.

previous
An option to specify whether to return the summary of previous states. Defaults to TRUE and only works if state = TRUE.

state_options
A list of threshold that would be used as state determination. See details for further information.

prefix
A character vector of length one that would be used as a prefix to the created bollinger bands columns. Defaults to "bbands".

h
A container for the names of "high". Leave to NULL as it will be populated by prep() function.

l
A container for the names of "low". Leave to NULL as it will be populated by prep() function.

c
A container for the names of "close". Leave to NULL as it will be populated by prep() function.

type
A container for the final series type that would be used ("hlc" or "c"). Leave to NULL as it will be populated by prep() function.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the created bollinger bands columns will be used as "predictors" in a model.

trained
A logical to indicate if the necessary informations for preprocessing have been estimated.

skip
A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_bbands object.

info
Options for tidy() method; whether to return tidied information for used "terms" or "params"

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

The output from this step are several new columns which contains the extracted Bollinger Bands features.

For basic output, this step will produces:

dn: lower band

ma: moving average

up: upper band

pctb: calculated

If state argument is TRUE, it will also produces:

state: current

state_count: cumulative count in current state

These states are determined using four different threshold, which listed in state_options arguments. These are the default threshold values:

high: pctb > high (the default is high = 1)

medhigh: high > pctb > medhigh (the default is medhigh = 0.75)

medlow: low < pctb < medlow (the default is medlow = 0.25)

low: pctb < low (the default is low = 0)

Note that the rest values would be categorized as "medium".

Additionally, if previous argument is TRUE, it will also provides some summary regarding previous Bollinger Bands states:

prev_state: previous state

prev_medium: previous medium-state

prev_break: previous break-state

Examples

# import libs
library(quantrecipes)

# an example recipes using built-in data
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_bbands(high, low, close,
    ma_fun = TTR::SMA,
    n = 20,
    sd_mult = 2
  ) %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 26,172 x 11
#>    datetime             open  high   low close volume turnover bbands_dn
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#>  1 2019-05-01 01:35:00 5340. 5340. 5331. 5331. 0.510     2722.     5322.
#>  2 2019-05-01 01:40:00 5338. 5339. 5333  5333  0.402     2146.     5324.
#>  3 2019-05-01 01:45:00 5335. 5337. 5333  5336. 0.819     4368.     5326.
#>  4 2019-05-01 01:50:00 5336. 5337. 5336. 5337. 0.0199     106.     5329.
#>  5 2019-05-01 01:55:00 5338. 5340. 5336. 5340. 1.39      7437.     5330.
#>  6 2019-05-01 02:00:00 5342. 5344. 5342  5344. 1.03      5504.     5330.
#>  7 2019-05-01 02:05:00 5344. 5352. 5344  5352. 1.26      6716.     5329.
#>  8 2019-05-01 02:10:00 5353  5355. 5348. 5348. 0.191     1020.     5328.
#>  9 2019-05-01 02:15:00 5349. 5369. 5348. 5362. 0.310     1661.     5326.
#> 10 2019-05-01 02:20:00 5361. 5362. 5355. 5355. 0.356     1908.     5325.
#> # … with 26,162 more rows, and 3 more variables: bbands_ma <dbl>,
#> #   bbands_up <dbl>, bbands_pctb <dbl>

# using state argument
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_bbands(high, low, close,
    ma_fun = TTR::SMA,
    n = 20,
    sd_mult = 2,
    state = TRUE
  ) %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 26,166 x 16
#>    datetime             open  high   low close volume turnover bbands_dn
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>  <dbl>    <dbl>     <dbl>
#>  1 2019-05-01 02:05:00 5344. 5352. 5344  5352. 1.26      6716.     5329.
#>  2 2019-05-01 02:10:00 5353  5355. 5348. 5348. 0.191     1020.     5328.
#>  3 2019-05-01 02:15:00 5349. 5369. 5348. 5362. 0.310     1661.     5326.
#>  4 2019-05-01 02:20:00 5361. 5362. 5355. 5355. 0.356     1908.     5325.
#>  5 2019-05-01 02:25:00 5355. 5362. 5355. 5360. 0.860     4610.     5325.
#>  6 2019-05-01 02:30:00 5362. 5364. 5360. 5364. 0.0272     146.     5325.
#>  7 2019-05-01 02:35:00 5360. 5360. 5358. 5360. 0.263     1411.     5325.
#>  8 2019-05-01 02:40:00 5360. 5361. 5358. 5359. 1.03      5499.     5326.
#>  9 2019-05-01 02:45:00 5359. 5361. 5357. 5359. 0.515     2759.     5326.
#> 10 2019-05-01 02:50:00 5358. 5359. 5357. 5357. 1.17      6252.     5327.
#> # … with 26,156 more rows, and 8 more variables: bbands_ma <dbl>,
#> #   bbands_up <dbl>, bbands_pctb <dbl>, bbands_state <fct>,
#> #   bbands_state_count <int>, bbands_prev_state <fct>,
#> #   bbands_prev_medium <fct>, bbands_prev_break <fct>