Calculate cumulative returns based on specified actions

step_cumret creates a specification of a recipe step that will calculate cumulative returns from a set of strategy actions and historical prices.

step_cumret(recipe, ..., prices, fee = 0.001, mult = FALSE,
  buy = "buy", sell = "sell", hold = "hold", prefix = "cumret",
  actions = NULL, role = "predictor", trained = FALSE,
  skip = FALSE, id = rand_id("cumret"))

# S3 method for step_cumret
tidy(x, info = "terms", ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections (from recipes package) for more details.

prices
A quoted column name which will be used as historical price data.

fee
A numeric vector pf length one, which will be used as fee for each trade action.

mult
A logical vector to specify whether to return the cumulative returns as cash multiplier or just the return. Defaults to FALSE

buy
A character vector of length one which indicates buy action(s) inside the selected actions columns. Default to "buy".

sell
A character vector of length one which indicates sell action(s) inside the selected actions columns. Default to "sell".

hold
A character vector of length one which indicates hold action(s) inside the selected actions columns. Default to "hold".

prefix
A character vector of length one that would be used as a prefix to the created columns.

actions
A container for selected actions columns. Leave to NULL as it will be populated by prep() function.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the created columns will be used as "predictors" in a model.

trained
A logical to indicate if the necessary informations for preprocessing have been estimated.

skip
A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_cumret object.

info
Options for tidy() method; whether to return tidied information for used "terms" or "params"

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

This step will return the calculated (log) cumulative returns, either just as a return, or as a cash multiplier for each selected columns.

Examples

# import libs
library(quantrecipes)

# basic usage
rec <- recipe(. ~ ., data = actions) %>%
  step_cumret(benchmark, portfolio, prices = "close") %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 863 x 6
#>    datetime            close benchmark portfolio cumret_benchmark
#>    <dttm>              <dbl> <fct>     <fct>                <dbl>
#>  1 2019-07-27 23:55:00 9486. buy       buy               -0.00405
#>  2 2019-07-28 00:00:00 9457. hold      hold              -0.0104 
#>  3 2019-07-28 00:05:00 9396. hold      hold              -0.00879
#>  4 2019-07-28 00:10:00 9412. hold      hold              -0.00559
#>  5 2019-07-28 00:15:00 9442. hold      hold              -0.00726
#>  6 2019-07-28 00:20:00 9426. hold      hold              -0.00533
#>  7 2019-07-28 00:25:00 9445. hold      hold              -0.00482
#>  8 2019-07-28 00:30:00 9450. hold      hold              -0.00458
#>  9 2019-07-28 00:35:00 9452. hold      hold              -0.00732
#> 10 2019-07-28 00:40:00 9426. hold      hold              -0.00715
#> # … with 853 more rows, and 1 more variable: cumret_portfolio <dbl>