Extract moving average features

step_ma creates a specification of a recipe step that will extract moving average features from an asset price historical data.

step_ma(recipe, ..., weights = NULL, ma_fun = TTR::SMA, n = 10,
  ma_options = list(), state = FALSE, ratio = TRUE, prefix = "ma",
  prices = NULL, role = "predictor", trained = FALSE, skip = FALSE,
  id = rand_id("ma"))

# S3 method for step_ma
tidy(x, info = "terms", ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections (from recipes package) for more details.

weights
A character vector of length one that specify a column name, or a numeric vector for ma_fun that has wts or volume argument. See details for more information.

ma_fun
A function to extract moving average, or a character vector of length one which specify a moving average function. Defaults to TTR::SMA.

n
A numeric vector of length one which specify the moving average window. The default is 10.

ma_options
A list of additional argument(s) that would be passed to ma_fun function.

state
An option to specify whether to return the current states of the calculated moving averages. See details for more informations. Defaults to TRUE.

ratio
Whether to return the moving average spread as ratio or absolute difference. See details for more informations. Defaults to TRUE.

prefix
A character vector of length one that would be used as a prefix to the created moving average columns. Defaults to "ma".

prices
A container for asd. Leave to NULL as it will be populated by prep() function.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the created moving average columns will be used as "predictors" in a model.

trained
A logical to indicate if the necessary informations for preprocessing have been estimated.

skip
A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_ma object.

info
Options for tidy() method; whether to return tidied information for used "terms" or "params"

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

The output from this step are several new columns which contains the extracted moving average features.

For basic output, this step will produces:

value: the moving average value

If state argument is TRUE, it will also produces:

spread: current spread to the prices variable

state: current state of the spread; "bullish" or "bearish"

Note that if ratio argument is TRUE, the spread value would be returned as a ratio to prices variable instead of an absolute difference.

Examples

# import libs
library(quantrecipes)

# an example recipes using built-in data
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_ma(close, ma_fun = TTR::SMA, n = 10) %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 26,182 x 8
#>    datetime             open  high   low close   volume turnover ma_close_value
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>          <dbl>
#>  1 2019-05-01 00:45:00 5336. 5336. 5331. 5331. 0.165      881.            5334.
#>  2 2019-05-01 00:50:00 5336. 5336. 5333  5335. 0.141      752.            5336.
#>  3 2019-05-01 00:55:00 5333. 5334. 5331. 5331. 0.000811     4.32          5337.
#>  4 2019-05-01 01:00:00 5337. 5338. 5337. 5338. 0.400     2137.            5338.
#>  5 2019-05-01 01:05:00 5341  5346. 5336. 5342. 0.583     3116.            5339.
#>  6 2019-05-01 01:10:00 5342. 5343. 5339. 5339. 3.10     16549.            5338.
#>  7 2019-05-01 01:15:00 5339. 5346. 5339. 5346. 0.182      972.            5338.
#>  8 2019-05-01 01:20:00 5347  5349. 5340. 5340. 1.05      5623.            5337.
#>  9 2019-05-01 01:25:00 5339. 5339. 5330  5332. 0.755     4030.            5337.
#> 10 2019-05-01 01:30:00 5331. 5338  5331  5338  0.963     5136.            5337.
#> # … with 26,172 more rows

# using state argument
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_ma(close, ma_fun = TTR::SMA, n = 10, state = TRUE) %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 26,182 x 10
#>    datetime             open  high   low close  volume turnover ma_close_value
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>   <dbl>    <dbl>          <dbl>
#>  1 2019-05-01 00:45:00 5336. 5336. 5331. 5331. 1.65e-1   881.            5334.
#>  2 2019-05-01 00:50:00 5336. 5336. 5333  5335. 1.41e-1   752.            5336.
#>  3 2019-05-01 00:55:00 5333. 5334. 5331. 5331. 8.11e-4     4.32          5337.
#>  4 2019-05-01 01:00:00 5337. 5338. 5337. 5338. 4.00e-1  2137.            5338.
#>  5 2019-05-01 01:05:00 5341  5346. 5336. 5342. 5.83e-1  3116.            5339.
#>  6 2019-05-01 01:10:00 5342. 5343. 5339. 5339. 3.10e+0 16549.            5338.
#>  7 2019-05-01 01:15:00 5339. 5346. 5339. 5346. 1.82e-1   972.            5338.
#>  8 2019-05-01 01:20:00 5347  5349. 5340. 5340. 1.05e+0  5623.            5337.
#>  9 2019-05-01 01:25:00 5339. 5339. 5330  5332. 7.55e-1  4030.            5337.
#> 10 2019-05-01 01:30:00 5331. 5338  5331  5338  9.63e-1  5136.            5337.
#> # … with 26,172 more rows, and 2 more variables: ma_close_spread <dbl>,
#> #   ma_close_state <fct>

# using pass-through options
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_ma(close,
    ma_fun = TTR::EMA,
    n = 10,
    ma_options = list(wilder = TRUE),
    state = TRUE
  ) %>%
  step_naomit(all_predictors()) %>%
  prep()

# get preprocessed data
juice(rec)

#> # A tibble: 26,182 x 10
#>    datetime             open  high   low close  volume turnover ma_close_value
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>   <dbl>    <dbl>          <dbl>
#>  1 2019-05-01 00:45:00 5336. 5336. 5331. 5331. 1.65e-1   881.            5334.
#>  2 2019-05-01 00:50:00 5336. 5336. 5333  5335. 1.41e-1   752.            5335.
#>  3 2019-05-01 00:55:00 5333. 5334. 5331. 5331. 8.11e-4     4.32          5334.
#>  4 2019-05-01 01:00:00 5337. 5338. 5337. 5338. 4.00e-1  2137.            5335.
#>  5 2019-05-01 01:05:00 5341  5346. 5336. 5342. 5.83e-1  3116.            5335.
#>  6 2019-05-01 01:10:00 5342. 5343. 5339. 5339. 3.10e+0 16549.            5336.
#>  7 2019-05-01 01:15:00 5339. 5346. 5339. 5346. 1.82e-1   972.            5337.
#>  8 2019-05-01 01:20:00 5347  5349. 5340. 5340. 1.05e+0  5623.            5337.
#>  9 2019-05-01 01:25:00 5339. 5339. 5330  5332. 7.55e-1  4030.            5336.
#> 10 2019-05-01 01:30:00 5331. 5338  5331  5338  9.63e-1  5136.            5337.
#> # … with 26,172 more rows, and 2 more variables: ma_close_spread <dbl>,
#> #   ma_close_state <fct>

# using volume based moving average
rec <- recipe(. ~ ., data = btcusdt) %>%
  step_ma(close,
    weights = "volume",
    ma_fun = TTR::VWMA,
    n = 10,
    state = TRUE
  ) %>%
  step_naomit(all_predictors()) %>%
  prep()

#> Warning: the condition has length > 1 and only the first element will be used

# get preprocessed data
juice(rec)

#> # A tibble: 26,182 x 10
#>    datetime             open  high   low close  volume turnover ma_close_value
#>    <dttm>              <dbl> <dbl> <dbl> <dbl>   <dbl>    <dbl>          <dbl>
#>  1 2019-05-01 00:45:00 5336. 5336. 5331. 5331. 1.65e-1   881.            5339.
#>  2 2019-05-01 00:50:00 5336. 5336. 5333  5335. 1.41e-1   752.            5341.
#>  3 2019-05-01 00:55:00 5333. 5334. 5331. 5331. 8.11e-4     4.32          5341.
#>  4 2019-05-01 01:00:00 5337. 5338. 5337. 5338. 4.00e-1  2137.            5341.
#>  5 2019-05-01 01:05:00 5341  5346. 5336. 5342. 5.83e-1  3116.            5342.
#>  6 2019-05-01 01:10:00 5342. 5343. 5339. 5339. 3.10e+0 16549.            5341.
#>  7 2019-05-01 01:15:00 5339. 5346. 5339. 5346. 1.82e-1   972.            5340.
#>  8 2019-05-01 01:20:00 5347  5349. 5340. 5340. 1.05e+0  5623.            5339.
#>  9 2019-05-01 01:25:00 5339. 5339. 5330  5332. 7.55e-1  4030.            5338.
#> 10 2019-05-01 01:30:00 5331. 5338  5331  5338  9.63e-1  5136.            5338.
#> # … with 26,172 more rows, and 2 more variables: ma_close_spread <dbl>,
#> #   ma_close_state <fct>